+++ sonar-polymorphic-views/src/main/java/be/kuleuven/cs/oss/resourcevisualizations/ResourceVisualization.java
@@ -22,9 +22,9 @@ public interface ResourceVisualization {
 	
 	
 	
-	public double getWidth();
+	public int getWidth();
 	
-	public double getHeight();
+	public int getHeight();
 	
 	/**
 	 * Gets the Size of this RV
